refactor: consolidate calendar invitation handling
