+++ apps/web/utils/actions/calendar-invitation.ts
@@ -3,7 +3,7 @@
 import { actionClient } from "@/utils/actions/safe-action";
 import { respondToCalendarInvitationBody } from "@/utils/actions/calendar-invitation.validation";
 import { createEmailProvider } from "@/utils/email/provider";
-import { respondToCalendarInvitation } from "@/utils/calendar/respond-to-invitation";
+import { respondToCalendarInvitation } from "@/utils/calendar/invitations/service";
 
 export const respondToCalendarInvitationAction = actionClient
   .metadata({ name: "respondToCalendarInvitation" })
