+++ apps/web/utils/calendar/constants.ts
@@ -1,7 +1,2 @@
 export const CALENDAR_STATE_COOKIE_NAME = "calendar_state";
 export const CALENDAR_ONBOARDING_RETURN_COOKIE = "calendar_onboarding_return";
-
-export const CALENDAR_INVITATION_LIMITS = {
-  content: 256_000,
-  encoded: 350_000,
-} as const;
