+++ apps/web/utils/calendar/event-provider.ts
@@ -1,9 +1,13 @@
+import { SafeError } from "@/utils/error";
 import prisma from "@/utils/prisma";
 import type { Logger } from "@/utils/logger";
 import type { CalendarEventProvider } from "@/utils/calendar/event-types";
 import { GoogleCalendarEventProvider } from "@/utils/calendar/providers/google-events";
 import { MicrosoftCalendarEventProvider } from "@/utils/calendar/providers/microsoft-events";
-import { isGoogleProvider } from "@/utils/email/provider-types";
+import {
+  isGoogleProvider,
+  isMicrosoftProvider,
+} from "@/utils/email/provider-types";
 
 /**
  * Create calendar event providers for all connected calendars.
@@ -38,32 +42,14 @@ export async function createCalendarEventProviders(
     if (!connection.refreshToken) continue;
 
     try {
-      if (isGoogleProvider(connection.provider)) {
-        providers.push(
-          new GoogleCalendarEventProvider(
-            {
-              accessToken: connection.accessToken,
-              connectionId: connection.id,
-              refreshToken: connection.refreshToken,
-              expiresAt: connection.expiresAt?.getTime() ?? null,
-              emailAccountId,
-            },
-            logger,
-          ),
-        );
-      } else if (connection.provider === "microsoft") {
-        providers.push(
-          new MicrosoftCalendarEventProvider(
-            {
-              accessToken: connection.accessToken,
-              refreshToken: connection.refreshToken,
-              expiresAt: connection.expiresAt?.getTime() ?? null,
-              emailAccountId,
-            },
-            logger,
-          ),
-        );
-      }
+      if (
+        !isGoogleProvider(connection.provider) &&
+        !isMicrosoftProvider(connection.provider)
+      )
+        continue;
+      providers.push(
+        createCalendarEventProvider({ connection, emailAccountId, logger }),
+      );
     } catch (error) {
       logger.error("Failed to create calendar event provider", {
         provider: connection.provider,
@@ -73,4 +59,38 @@ export async function createCalendarEventProviders(
   }
 
   return providers;
+}
+
+export function createCalendarEventProvider({
+  connection,
+  emailAccountId,
+  logger,
+}: {
+  connection: {
+    accessToken: string | null;
+    expiresAt: Date | null;
+    id: string;
+    provider: string;
+    refreshToken: string | null;
+  };
+  emailAccountId: string;
+  logger: Logger;
+}) {
+  const providerParams = {
+    accessToken: connection.accessToken,
+    connectionId: connection.id,
+    refreshToken: connection.refreshToken,
+    expiresAt: connection.expiresAt?.getTime() ?? null,
+    emailAccountId,
+  };
+
+  if (isGoogleProvider(connection.provider)) {
+    return new GoogleCalendarEventProvider(providerParams, logger);
+  }
+
+  if (isMicrosoftProvider(connection.provider)) {
+    return new MicrosoftCalendarEventProvider(providerParams, logger);
+  }
+
+  throw new SafeError("Unsupported calendar provider");
 }
