+++ apps/web/utils/calendar/invitations/constants.ts
@@ -1,0 +1,4 @@
+export const CALENDAR_INVITATION_LIMITS = {
+  content: 256_000,
+  encoded: 350_000,
+} as const;
